pyproject.toml
[tool.poetry]
name = "libzfseasy"
version = "0.1.0"
description = "Python interface to the ZFS utilities"
authors = ["Radoslaw Kamil Ejsmont <radoslaw@ejsmont.net>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
# libzfs = {git = "https://gitlab.vibfolks.eu/rejsmont/py-libzfs.git"}

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
notebook = "^7.5.0"
pytest = "^9.0.2"
pytest-mock = "^3.12.0"
pytest-cov = "^7.0.0"
pytest-xdist = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
